[Silver I] Title: Subway, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/21959. Subway/Subway.cc
@@ -15,7 +15,7 @@ using namespace std;
 
 typedef unsigned long long ull;
 
-ull inline nC2(int n)
+ull inline nC2(ull n)
 {
 	return n * (n - 1) / 2;
 }
